+++ archive/bohartadams_fit.tikz
@@ -1,0 +1,134 @@
+% This file was created by matplotlib2tikz v0.6.15.
+\begin{tikzpicture}
+
+\begin{axis}[
+xlabel={Time (s)},
+ylabel={Concentration (mg/mL)},
+xmin=-75, xmax=1575,
+ymin=2.92762715543126, ymax=5.04339913394363,
+width=12cm,
+height=8cm,
+tick pos=left,
+xmajorgrids,
+x grid style={white!69.80392156862744!black},
+ymajorgrids,
+y grid style={white!69.80392156862744!black},
+axis line style={white!73.72549019607844!black},
+axis background/.style={fill=white!93.33333333333333!black},
+legend cell align={left},
+legend style={at={(0.97,0.03)}, anchor=south east, draw=white!80.0!black, fill=white!93.33333333333333!black},
+legend entries={{Experiment},{Fit}}
+]
+\addlegendimage{no markers, black}
+\addlegendimage{no markers, black}
+\addplot [thick, black, mark=*, mark size=3, mark options={solid,fill=none}, only marks]
+table {%
+0 3.023798609
+30 3.114830065
+60 3.193132997
+90 3.344098032
+120 3.545952177
+150 3.71131258
+180 3.880179942
+210 4.051567078
+240 4.145738602
+270 4.204649925
+300 4.310641861
+330 4.373957515
+360 4.449874592
+390 4.523475265
+420 4.557658434
+450 4.621247387
+480 4.665646744
+510 4.688958049
+540 4.724696445
+570 4.763434982
+600 4.777761936
+630 4.817210579
+660 4.840685558
+690 4.846937061
+720 4.861563492
+750 4.862069321
+780 4.865633011
+810 4.883685875
+840 4.888158989
+870 4.893807769
+900 4.89773283
+930 4.902150154
+960 4.908982992
+990 4.91435442
+1020 4.919867516
+1050 4.916160679
+1080 4.928572059
+1110 4.917313385
+1140 4.922553825
+1170 4.919561267
+1200 4.92966497
+1230 4.934463024
+1260 4.935445404
+1290 4.937192059
+1320 4.930878401
+1350 4.932299042
+1380 4.933749437
+1410 4.929975319
+1440 4.935555362
+1470 4.937137485
+1500 4.940053844
+};
+\addplot [thick, black, dashed]
+table {%
+0 3.023798609
+30 3.19016829228515
+60 3.34884542927062
+90 3.49899362818041
+120 3.64000586564161
+150 3.77150029319125
+180 3.89330564495151
+210 4.0054390307107
+240 4.10807895432934
+270 4.2015359857764
+300 4.28622310586129
+330 4.36262796484208
+360 4.43128751880771
+390 4.49276612448273
+420 4.54763743012845
+450 4.59646968762859
+480 4.63981459908541
+510 4.67819914319466
+540 4.71211996392947
+570 4.74204002945974
+600 4.76838672953713
+630 4.79155159756503
+660 4.81189092720335
+690 4.82972700128801
+720 4.84534995243694
+750 4.85901985152631
+780 4.87096894278508
+810 4.88140399453553
+840 4.89050857217624
+870 4.89844525658941
+900 4.90535774202648
+930 4.91137281740927
+960 4.91660216943786
+990 4.92114404916631
+1020 4.92508478844966
+1050 4.9285001708334
+1080 4.93145665700925
+1110 4.93401248554476
+1140 4.93621865975246
+1170 4.93811981662907
+1200 4.93975500182549
+1230 4.94115835371013
+1260 4.94235970982919
+1290 4.94338513732543
+1320 4.94425740269117
+1350 4.9449963852399
+1380 4.9456194355895
+1410 4.94614169962278
+1440 4.94657638949651
+1470 4.94693503171005
+1500 4.94722768037489
+};
+\end{axis}
+
+\end{tikzpicture}
