+++ archive/lmfit_notstrict.tikz
