+++ archive/thomas_fit.tikz
